+++ Plutarch/TryFrom.hs
@@ -15,14 +15,11 @@ module Plutarch.TryFrom (
 
 import Data.Kind (Constraint)
 import Data.Proxy (Proxy (Proxy))
-
+import GHC.TypeLits (ErrorMessage (ShowType, Text, (:<>:)), TypeError)
 import Plutarch.Internal (PType, Term, punsafeCoerce)
 import Plutarch.Internal.PlutusType (PContravariant, PCovariant, PInner)
 import Plutarch.Internal.Witness (witness)
-
 import Plutarch.Reducible (Reduce)
-
-import GHC.TypeLits (ErrorMessage (ShowType, Text, (:<>:)), TypeError)
 
 data PSubtypeRelation
   = PSubtypeRelation