+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Scripts.hs
@@ -9,6 +9,7 @@ module Plutarch.LedgerApi.Scripts (
   PRedeemerHash (..),
 ) where
 
+import Plutarch.Builtin (PDataNewtype (PDataNewtype))
 import Plutarch.LedgerApi.Utils (Mret)
 import Plutarch.Lift (
   DerivePConstantViaBuiltin (DerivePConstantViaBuiltin),
@@ -21,7 +22,7 @@ import Plutarch.Unsafe (punsafeCoerce)
 import PlutusLedgerApi.V3 qualified as Plutus
 
 -- | @since 2.0.0
-newtype PScriptHash (s :: S) = PScriptHash (Term s PByteString)
+newtype PScriptHash (s :: S) = PScriptHash (Term s (PDataNewtype PByteString))
   deriving stock
     ( -- | @since 2.0.0
       Generic
@@ -55,6 +56,18 @@ deriving via
   instance
     PConstantDecl Plutus.ScriptHash
 
+-- | @since 3.1.0
+instance PTryFrom PData PScriptHash where
+  type PTryFromExcess PData PScriptHash = Mret PScriptHash
+  ptryFrom' opq = runTermCont $ do
+    unwrapped <- tcont . plet $ ptryFrom @(PAsData PByteString) opq snd
+    tcont $ \f ->
+      pif
+        (plengthBS # unwrapped #== 28)
+        (f ())
+        (ptraceInfoError "ptryFrom(PScriptHash): must be 28 bytes long")
+    pure (punsafeCoerce opq, pcon . PScriptHash . pcon . PDataNewtype . pdata $ unwrapped)
+
 -- | @since 2.0.0
 instance PTryFrom PData (PAsData PScriptHash) where
   type PTryFromExcess PData (PAsData PScriptHash) = Mret PScriptHash
@@ -65,7 +78,7 @@ instance PTryFrom PData (PAsData PScriptHash) where
         (plengthBS # unwrapped #== 28)
         (f ())
         (ptraceInfoError "ptryFrom(PScriptHash): must be 28 bytes long")
-    pure (punsafeCoerce opq, pcon . PScriptHash $ unwrapped)
+    pure (punsafeCoerce opq, pcon . PScriptHash . pcon . PDataNewtype . pdata $ unwrapped)
 
 -- | @since 2.0.0
 newtype PDatum (s :: S) = PDatum (Term s PData)
@@ -82,6 +95,8 @@ newtype PDatum (s :: S) = PDatum (Term s PData)
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.1.0
+      PTryFrom PData
     )
 
 -- | @since 2.0.0
@@ -98,8 +113,11 @@ deriving via
   instance
     PConstantDecl Plutus.Datum
 
+-- | @since 3.1.0
+instance PTryFrom PData (PAsData PDatum)
+
 -- | @since 2.0.0
-newtype PDatumHash (s :: S) = PDatumHash (Term s PByteString)
+newtype PDatumHash (s :: S) = PDatumHash (Term s (PDataNewtype PByteString))
   deriving stock
     ( -- | @since 2.0.0
       Generic
@@ -117,6 +135,8 @@ newtype PDatumHash (s :: S) = PDatumHash (Term s PByteString)
       POrd
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.1.0
+      PTryFrom PData
     )
 
 -- | @since 2.0.0
@@ -133,6 +153,9 @@ deriving via
   instance
     PConstantDecl Plutus.DatumHash
 
+-- | @since 3.1.0
+instance PTryFrom PData (PAsData PDatumHash)
+
 -- | @since 2.0.0
 newtype PRedeemer (s :: S) = PRedeemer (Term s PData)
   deriving stock
@@ -148,6 +171,8 @@ newtype PRedeemer (s :: S) = PRedeemer (Term s PData)
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.1.0
+      PTryFrom PData
     )
 
 -- | @since 2.0.0
@@ -164,8 +189,11 @@ deriving via
   instance
     PConstantDecl Plutus.Redeemer
 
+-- | @since 3.1.0
+instance PTryFrom PData (PAsData PRedeemer)
+
 -- | @since 2.0.0
-newtype PRedeemerHash (s :: S) = PRedeemerHash (Term s PByteString)
+newtype PRedeemerHash (s :: S) = PRedeemerHash (Term s (PDataNewtype PByteString))
   deriving stock
     ( -- | @since 3.1.0
       Generic
@@ -183,6 +211,8 @@ newtype PRedeemerHash (s :: S) = PRedeemerHash (Term s PByteString)
       POrd
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.1.0
+      PTryFrom PData
     )
 
 -- | @since 3.1.0
@@ -198,3 +228,6 @@ deriving via
   (DerivePConstantViaBuiltin Plutus.RedeemerHash PRedeemerHash PByteString)
   instance
     PConstantDecl Plutus.RedeemerHash
+
+-- | @since 3.1.0
+instance PTryFrom PData (PAsData PRedeemerHash)