+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Time.hs
@@ -5,13 +5,14 @@ module Plutarch.LedgerApi.Time (
   PPosixTime (..),
 ) where
 
+import Plutarch.Builtin (PDataNewtype (PDataNewtype))
 import Plutarch.LedgerApi.Utils (Mret)
 import Plutarch.Lift (
   DerivePConstantViaNewtype (DerivePConstantViaNewtype),
   PConstantDecl,
   PUnsafeLiftDecl (PLifted),
  )
-import Plutarch.Num (PNum)
+import Plutarch.Num (PNum (pabs, pfromInteger, pnegate, psignum, (#*), (#+), (#-)))
 import Plutarch.Prelude
 import Plutarch.Reducible (Reduce)
 import Plutarch.TryFrom (PTryFrom (PTryFromExcess, ptryFrom'))
@@ -19,7 +20,7 @@ import Plutarch.Unsafe (punsafeCoerce)
 import PlutusLedgerApi.V3 qualified as Plutus
 
 -- | @since 2.0.0
-newtype PPosixTime (s :: S) = PPosixTime (Term s PInteger)
+newtype PPosixTime (s :: S) = PPosixTime (Term s (PDataNewtype PInteger))
   deriving stock
     ( -- | @since 2.0.0
       Generic
@@ -36,12 +37,73 @@ newtype PPosixTime (s :: S) = PPosixTime (Term s PInteger)
     , -- | @since 2.0.0
       POrd
     , -- | @since 2.0.0
-      PIntegral
-    , -- | @since 2.0.0
-      PNum
-    , -- | @since 2.0.0
       PShow
     )
+
+-- | @since 2.0.0
+instance PIntegral PPosixTime where
+  pdiv = phoistAcyclic $ plam $ \t1 t2 ->
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            pcon . PPosixTime . pcon . PDataNewtype . pdata $ pdiv # pfromData t1'' # pfromData t2''
+  pmod = phoistAcyclic $ plam $ \t1 t2 ->
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            pcon . PPosixTime . pcon . PDataNewtype . pdata $ pmod # pfromData t1'' # pfromData t2''
+  pquot = phoistAcyclic $ plam $ \t1 t2 ->
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            pcon . PPosixTime . pcon . PDataNewtype . pdata $ pquot # pfromData t1'' # pfromData t2''
+  prem = phoistAcyclic $ plam $ \t1 t2 ->
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            pcon . PPosixTime . pcon . PDataNewtype . pdata $ pquot # pfromData t1'' # pfromData t2''
+
+-- | @since 2.0.0
+instance PNum PPosixTime where
+  t1 #* t2 =
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            pcon . PPosixTime . pcon . PDataNewtype . pdata $ pfromData t1'' #+ pfromData t2''
+  t1 #+ t2 =
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            pcon . PPosixTime . pcon . PDataNewtype . pdata $ pfromData t1'' #+ pfromData t2''
+  t1 #- t2 =
+    pmatch t1 $ \(PPosixTime t1') ->
+      pmatch t2 $ \(PPosixTime t2') ->
+        pmatch t1' $ \(PDataNewtype t1'') ->
+          pmatch t2' $ \(PDataNewtype t2'') ->
+            plet (pfromData t1'' #- pfromData t2'') $ \res ->
+              pif
+                (0 #<= res)
+                (pcon . PPosixTime . pcon . PDataNewtype . pdata $ res)
+                (ptraceInfoError "PPosixTime subtraction gave a negative value")
+  pabs = phoistAcyclic $ plam id
+  pfromInteger i
+    | i < 0 = ptraceInfoError "Cannot make PPosixTime from a negative value"
+    | otherwise = pcon . PPosixTime . pcon . PDataNewtype . pdata . pconstant $ i
+  pnegate = phoistAcyclic $ plam $ \_ -> ptraceInfoError "PPosixTime can't be negative"
+  psignum = phoistAcyclic $ plam $ \t ->
+    pmatch t $ \(PPosixTime t') ->
+      pmatch t' $ \(PDataNewtype t'') ->
+        pcon . PPosixTime . pcon . PDataNewtype . pdata $
+          pif
+            (pfromData t'' #== 0)
+            0
+            1
 
 -- | @since 2.0.0
 instance DerivePlutusType PPosixTime where
@@ -57,6 +119,20 @@ deriving via
   instance
     PConstantDecl Plutus.POSIXTime
 
+-- | @since 3.1.0
+instance PTryFrom PData PPosixTime where
+  type PTryFromExcess PData PPosixTime = Mret PPosixTime
+  ptryFrom' ::
+    forall (s :: S) (r :: S -> Type).
+    Term s PData ->
+    ((Term s PPosixTime, Reduce (PTryFromExcess PData PPosixTime s)) -> Term s r) ->
+    Term s r
+  ptryFrom' opq = runTermCont $ do
+    (wrapped :: Term s (PAsData PInteger), unwrapped :: Term s PInteger) <-
+      tcont $ ptryFrom @(PAsData PInteger) opq
+    tcont $ \f -> pif (0 #<= unwrapped) (f ()) (ptraceInfoError "ptryFrom(POSIXTime): must be positive")
+    pure (punsafeCoerce wrapped, pcon . PPosixTime . pcon . PDataNewtype . pdata $ unwrapped)
+
 -- | @since 2.0.0
 instance PTryFrom PData (PAsData PPosixTime) where
   type PTryFromExcess PData (PAsData PPosixTime) = Mret PPosixTime
@@ -69,4 +145,4 @@ instance PTryFrom PData (PAsData PPosixTime) where
     (wrapped :: Term s (PAsData PInteger), unwrapped :: Term s PInteger) <-
       tcont $ ptryFrom @(PAsData PInteger) opq
     tcont $ \f -> pif (0 #<= unwrapped) (f ()) (ptraceInfoError "ptryFrom(POSIXTime): must be positive")
-    pure (punsafeCoerce wrapped, pcon $ PPosixTime unwrapped)
+    pure (punsafeCoerce wrapped, pcon . PPosixTime . pcon . PDataNewtype . pdata $ unwrapped)