+++ plutarch-ledger-api/CHANGELOG.md
@@ -26,7 +26,38 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
     * `PScriptInfo`
     * `PCommittee`
 * `PRationalData`, originally from `plutarch-extra`
-* `PTryFrom PData (PAsData PLovelace)` instance
+* `PTryFrom` instances:
+    * `PTryFrom PData (PAsData PLovelace)`
+    * `PTryFrom PData PCurrencySymbol`
+    * `PTryFrom PData PScriptHash`
+    * `PTryFrom PData PTxId`
+    * `PTryFrom PData (PAsData PTxId)`
+    * `PTryFrom PData PTxOutRef`
+    * `PTryFrom PData (PAsData PTxOutRef)`
+    * `PTryFrom PData PDatumHash` instance
+    * `PTryFrom PData (PAsData PDatumHash)`
+    * `PTryFrom PData POutputDatum`
+    * `PTryFrom PData (PAsData POutputDatum)`
+    * `PTryFrom PData PTxInfo`
+    * `PTryFrom PData (PAsData PTxInfo)`
+    * `PTryFrom PData a => PTryFrom PData (PInterval a)`
+    * `PTryFrom PData a => PTryFrom PData (PAsData (PInterval a))`
+    * `PTryFrom PData a => PTryFrom PData (PLowerBound a)`
+    * `PTryFrom PData a => PTryFrom PData (PAsData (PLowerBound a))`
+    * `PTryFrom PData a => PTryFrom PData (PUpperBound a)`
+    * `PTryFrom PData a => PTryFrom PData (PAsData (PUpperBound a))`
+    * `PTryFrom PData a => PTryFrom PData (PExtended a)`
+    * `PTryFrom PData a => PTryFrom PData (PAsData (PExtended a))`
+    * `PTryFrom PData PRedeemer`
+    * `PTryFrom PData (PAsData PRedeemer)`
+    * `PTryFrom PData PPosixTime`
+    * `PTryFrom PData PScriptContext`
+    * `PTryFrom PData (PAsData PScriptContext)`
+    * `PTryFrom PData PPubKeyHash`
+    * `PTryFrom PData PTokenName`
+* `PUnsafeLiftDecl (PExtended a)` instance
+* `PConstantDecl (Extended a)` instance
+* `PBoolData`, corresponding to PlutusTx's `Bool`
 
 ### Modified